use mmap from julia instead of numpy
On cobra, there was an error that the numpy mmaps can not be read anymore.
For this purpose I switch to julia mmaps. This resolved the issue.
However, there are potentially more:
1. the python function `var` is not called anymore. This means
that conversion (e.g. for e or uz) for variables that are not saved are not
applied! Also different mhd solver may require different conversions! I added
a function _var_from_patch that does the same thing, however only
for velocity and ee. Other may need to be added/or improved later!
This is not extensively tested! If there appear errors, dont forget this!