make conversions faster
most of the time was spent in EoS lookups, which is now much better.
A function was added that looks up in-place for dispatch (MUST.jl) lookup functions.
Also var_from_patch is inplace now, although the mmap still needs to be
replaced I think, because it is totally not needed. Not sure if it would be faster.
Currently ~40% of the time is spent reading the patch data from files and
~40% interpolating to the optical depth scale.
More timers were added.